move password hashing to Users model afterValidate hook + add validations
this change utilizes the sequelize afterValidate hook to overwrite the
incoming passwordHash value and replace it with a hashed version of the
password. using the afterValidate hook is beneficial because it allows me
to use sequelize validations on the plain text password before it gets
replaced by the hashed version of the password in the afterValidate hook.

using the built-in validation capabilities of sequelize means that it
will generate nice error messages that can be easily sent back to the
client.

this change creates a new issue related to the validation error
messages. specifically, the error message uses the db column name in the
message text sent back to the client. this column name (passwordHash) is
different than the property sent by the client (password), which could
cause confusion for the client. so the database column name should be
changed to "password" in a follow-up commit.