refactor PATCH /users so it creates a patch object
i am converting to using a patch object because i will conditionally add
values to the patch object only if they exist on the request body. i
will be adding extra properties that can be patched, so it is important
to build up a patch object so only the properties which are provided
will be updated.