return the user that was updated in PATCH /users response
previously, the PATCH /users response was returning the number of rows
affected by the users.update call. this info is not relevant for the
client and is also an atypical response for PATCH requests. By
convention, PATCH requests should return the entity which was patched,
in this case that is the user.