use transaction for DELETE /users + basic error handling
a transaction ensures that either the three destroy calls are processed
successfully or not at all. this should prevent the data from being in a
partial deleted state (such as likes and messages deleted, but user not
deleted) due to network issues or api server or db server crashing
midway through.

also adds basic error handling by returning a 500 response in case of
error.