remove classMethods.associate from models, define them in index.js
previously, the code in index.js was rather complicated. it dynamically
created an array of model names then called each model class method
called associate, passing to it the object (db) that contained all the
other models. this code was deleted in favor of a simpler approach where
the commands in each model's associate method was simply moved into
index.js.