add ON DELETE CASCADE to all foreign key constraints
this change causes messages and likes to automatically be deleted when a
user is deleted. it also causes all likes associated with a message to
be deleted when a message is deleted.

since the db is now handling the ON DELETE CASCADE, it means that we can
remove some of the code associated with manually deleting likes first
(in the case of DELETE /messages/:id) and with manually deleting likes
and messages (in the case of DELETE /users).