move swagger comments into specification.yaml and update to openapi 3
this change removes all @swagger comments found in the individual
controllers files and instead uses a separate specification.yaml.
because the project now uses a separate yaml file, it means that package
swagger-jsdoc is no longer needed.

while putting swagger spec details as comments next to the endpoint is a
nice idea, it does not work well in practice due to lack of
autoindenting and lack of linting in comments. it is much more helpful
to have these features by putting the spec in its own yaml file.

this commit also cleans up the code, deleting unecessary comments and
code as well as uninstall mocha and supertest (not being used).