update postinstall script to source .env file
the postinstall script would error out if a DATABASE_URL env variable
was not defined. it is assumed that this variable will already be
defined in a `.env` file, so that file is sourced in order to get the
DATABASE_URL env variable. This prevents having to include DATABASE_URL
definition with the `npm install` command.