change how non-200 responses are handled in controllers
this change gets rid of any res.send which included a non-200 status in
controller handlers. It replaces them by instead passing the error
object to `next` so that the top-level error handler can handle sending
the response.

this change also updates the specification.yaml so that the
aforementioned non-200 responses are included in the spec.

finally, the login handler has been updated to send back a 400 status
code if the username or password is invalid. this is a more proper
status code than 401 because a 401 is strictly used to indicate that
credentials must be provided but were not. In this case, credentials
(aka an api key) are not required but a valid username/password
properties are.