update authentication strategy to handle more error conditions
this change adds operation error handling for when the user does not
exist. this situation could happen if the user deletes their account and
then tries to use their previous valid token to try and perform another
auth-related request.

this change also adds error handling for when communication with the
database errors. In that case, the promise for User.findById will reject
and cause execution to jump to the catch block. The catch block simply
forwards whatever error that was produced into the done function. The
done function will pass any errors onto the express error handling
middlewares--which handes any 500 statuses.

this change also removes dependency on VError, instead opting for using
Object.assign to construct uniqueConstraintError. This error occurs when
the username is already taken in the database.