remap PATCH /users & DELETE /users & PUT /users/picture
in order to follow more semantic RESTful order, modifications to a user
are now nested under the /users/{userId} path. This ensures that all
user modification endpoints are nested in the /users collection path.

Additionally, this will allow for the potential to create admin users
which can operate on modifying other users.

this change also updates the postman collection to account for endpoint
remappings. moving these endpoints to a new path also required adding
parameter to include the userId in the each of the endpoints.