add pictureLocation to User model
this additional property will display the path to GET the user's picture
iff the user has a picture. when the user does not have a picture,
pictureLocation will have a value of null.

this change should make it easier for client code to know whether or not
it should create an img tag to display the user's profile picture or to
instead render a default profile picture (in the case when
pictureLocation is null).

todo: will try to come back and refactor this code so it is a little
cleaner. i think there is a better way to setup this psuedo property,
pictureLocation, with sequelize features. i will have to study it more.
in the meantime this code works.