change User model afterValidate hook to beforeUpdate hook
the afterValidate hook was running twice when updating a User instance.
this was causing the password to be double hashed. this was obviously a
problem for logging back in because login only dehashes once.

by changing this to the beforeUpdate hook, it does not cause validation
to run a second time.