fix bug where could not login again after updating user info
this bug was caused by the beforeUpdate hook in the User model. it was
only checking if user.password exists before hashing its value. this was
a problem when updating user details because the user instance ALWAYS
contains the user.password property. in the case when an updated
password was provided by the client, this code would hash that plain
text password and work fine. However, in the case when no updated
password was provided (like when just the about or displayName is being
updated by the user) then the if check was not strict enough, and it
would hash the existing hashed password. this caused subsequent logins
to no longer work.

the fix makes this if check more strict so that it will only hash the
password on the user instance when that password is NOT 60 characters
long (60 characters is a distinguishing mark of a hashed password
according to bcrypt spec). Since the user cannot provide an updated
plain text password longer than 20 characters, the 60 character length
is a suitable distinguishing mark of an already hashed password.