create migration to make username the primary key of users table
this commit also includes changing sequelize models to reflect this
change in the database users table. the Message and Like model for
sequelize now include the username attribute so the User model does not
need to be included in queries for messages and likes. This also
simplifies the getRawMessage and getRawLike helper functions because
they no longer have to get dataValues for the User model's username.

this commit also changes all controller endpoints which used User id or
userId to username instead.

this commit also updates the specification.yaml schemas from User id or
userId to username.

this commit also goes ahead and updates all sequelize Model.findById
calls to Model.findByPk calls (this method is more semantically correct
because some tables do not use an id for the primary key, as is the case
for the users table using username now).